refactor: new code directory structure
