+++ README.md
@@ -2,7 +2,7 @@
 
 This repository aims to build a continuously updating database of `sorry` statements in public Lean4 repositories. The idea is to use this as a basis for a continuously running benchmark which tests the performance of automated proof systems against *real world* Lean statements.
 
-For a detailed explanation of the project's motivation, philosophy, and goals, see [ABOUT.md](ABOUT.md).
+For a detailed explanation of the project's motivation, philosophy, and goals, see [doc/ABOUT.md](ABOUT.md).
 
 Currently we are building:
 
@@ -28,12 +28,12 @@ Scripts can be run from poetry's virtual environment by running
 To initialize a database file, one needs a json with a list of repositories to
 monitor. The folder `repo_lists` provides some examples. Then run for example
 
-`poetry run src/sorrydb/scripts/init_db.py --repos-file repo_lists/mock_repos.json --database-file mock_db.json`
+`poetry run src/scripts/init_db.py --repos-file data/repo_lists/mock_repos.json --database-file mock_db.json`
 
 This provides an initialised database `mock_db.json` which does not yet contain
 any sorries. Now one can update the database repeatedly using:
 
-`poetry run src/sorrydb/scripts/update_db.py --database-file mock_db.json`
+`poetry run src/scripts/update_db.py --database-file mock_db.json`
 
 
 ## Contributing
