--- repo_lists/active_repos_without_mathlib.json
@@ -1,83 +1,0 @@
-{
-    "documentation": "List of active repositories pulled from reservoir, each has been recently updated in the last 5 days (as of 2025-03-09), and has at least 10 GitHub stars, minus mathlib4",
-    "repos": [
-        {
-            "remote": "https://github.com/m4lvin/pdl"
-        },
-        {
-            "remote": "https://github.com/vihdzp/combinatorial-games"
-        },
-        {
-            "remote": "https://github.com/leanprover-community/flt-regular"
-        },
-        {
-            "remote": "https://github.com/dwrensha/compfiles"
-        },
-        {
-            "remote": "https://github.com/emilyriehl/InfinityCosmos"
-        },
-        {
-            "remote": "https://github.com/HEPLean/PhysLean"
-        },
-        {
-            "remote": "https://github.com/FormalizedFormalLogic/arithmetization"
-        },
-        {
-            "remote": "https://github.com/FormalizedFormalLogic/Foundation"
-        },
-        {
-            "remote": "https://github.com/forked-from-1kasper/GroundZero"
-        },
-        {
-            "remote": "https://github.com/leanprover-community/SphereEversion"
-        },
-        {
-            "remote": "https://github.com/ah1112/synthetic_euclid_4"
-        },
-        {
-            "remote": "https://github.com/leanprover/TensorLib"
-        },
-        {
-            "remote": "https://github.com/Ivan-Sergeyev/Seymour"
-        },
-        {
-            "remote": "https://github.com/vltanh/lean4-analysis-tao"
-        },
-        {
-            "remote": "https://github.com/fpvandoorn/carleson"
-        },
-        {
-            "remote": "https://github.com/teorth/PFR"
-        },
-        {
-            "remote": "https://github.com/Verified-zkEVM/Zklib"
-        },
-        {
-            "remote": "https://github.com/YaelDillies/LeanCamCombi"
-        },
-        {
-            "remote": "https://github.com/YaelDillies/LeanAPAP"
-        },
-        {
-            "remote": "https://github.com/ImperialCollegeLondon/FLT"
-        },
-        {
-            "remote": "https://github.com/teorth/expdb"
-        },
-        {
-            "remote": "https://github.com/optsuite/convex"
-        },
-        {
-            "remote": "https://github.com/leanprover-community/ConNF"
-        },
-        {
-            "remote": "https://github.com/Timeroot/computableReal"
-        },
-        {
-            "remote": "https://github.com/ImperialCollegeLondon/M1F-explained"
-        },
-        {
-            "remote": "https://github.com/mo271/FormalBook"
-        }
-    ]
-}
