--- repo_lists/mock_repos.json
@@ -1,11 +1,0 @@
-{
-    "documentation": "Mock repositories for testing.",
-    "repos": [
-        {
-            "remote": "https://github.com/austinletson/sorryClientTestRepo"
-        },
-        {
-            "remote": "https://github.com/austinletson/sorryClientTestRepoMath" 
-        }
-    ]
-}
