--- repo_lists/mock_single.json
@@ -1,8 +1,0 @@
-{
-    "documentation": "Smallest unit for faster testing.",
-    "repos": [
-        {
-            "remote": "https://github.com/austinletson/sorryClientTestRepo"
-        }
-    ]
-}
