--- repo_lists/mock_with_carleson.json
@@ -1,14 +1,0 @@
-{
-    "documentation": "Mock repositories plus Carleson, for testing.",
-    "repos": [
-        {
-            "remote": "https://github.com/austinletson/sorryClientTestRepo"
-        },
-        {
-            "remote": "https://github.com/austinletson/sorryClientTestRepoMath" 
-        },
-        {
-            "remote": "https://github.com/fpvandoorn/carleson"
-        }
-    ]
-}
