--- repo_lists/mock_with_mathlib.json
@@ -1,14 +1,0 @@
-{
-    "documentation": "Mock repositories plus Mathlib, for testing.",
-    "repos": [
-        {
-            "remote": "https://github.com/austinletson/sorryClientTestRepo"
-        },
-        {
-            "remote": "https://github.com/austinletson/sorryClientTestRepoMath" 
-        },
-        {
-            "remote": "https://github.com/leanprover-community/mathlib4"
-        }
-    ]
-}
