--- sorry_samples/README.md
@@ -1,10 +1,0 @@
-# Sorry Samples
-
-This directory contains sample JSON files representing individual "sorry" instances for testing purposes. Each JSON file contains exactly one sorry with all the information needed to locally reproduce it.
-
-## Purpose
-
-These samples define the precise interface between:
-
-- The **server** (leaderboard) that provides sorry challenges
-- The **client** (theorem proving model) that attempts to solve them. The output of the client should be a single string, replacing the original "sorry".
