+++ src/sorrydb/build_database.py
@@ -8,8 +8,8 @@ import uuid
 from pathlib import Path
 from typing import Optional
 
-from sorrydb.crawler.git_ops import leaf_commits, remote_heads_hash
-from sorrydb.database.process_sorries import prepare_and_process_lean_repo
+from utils.git_ops import leaf_commits, remote_heads_hash
+from sorrydb.process_sorries import prepare_and_process_lean_repo
 
 # Create a module-level logger
 logger = logging.getLogger(__name__)
